pyproject.toml
[project]
name = "smm-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "apscheduler==3.10.4",
    "cerebras-cloud-sdk==1.14.0",
    "flask==3.0.3",
    "flask-caching==2.3.0",
    "flask-cors==4.0.1",
    "google-auth==2.29.0",
    "google-cloud-aiplatform==1.52.0",
    "gspread==6.1.2",
    "pillow>=10.3.0",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
]
